+++ src/CommandLineUtils/CommandLineApplication{T}.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.ComponentModel;
 using McMaster.Extensions.CommandLineUtils.Abstractions;
 using McMaster.Extensions.CommandLineUtils.Conventions;
 using McMaster.Extensions.CommandLineUtils.HelpText;
@@ -19,20 +20,17 @@ namespace McMaster.Extensions.CommandLineUtils
         private Func<TModel> _modelFactory = DefaultModelFactory;
 
 #nullable disable
-#pragma warning disable RS0027 // Public API with optional parameter(s) should have the most parameters amongst its public overloads.
         /// <summary>
-        /// Initializes a new instance of <see cref="CommandLineApplication"/>.
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
         /// </summary>
-        /// <param name="throwOnUnexpectedArg">Initial value for <see cref="CommandLineApplication.ThrowOnUnexpectedArgument"/>.</param>
-        public CommandLineApplication(bool throwOnUnexpectedArg = true)
-#pragma warning restore RS0027 // Public API with optional parameter(s) should have the most parameters amongst its public overloads.
-            : base(throwOnUnexpectedArg)
+        public CommandLineApplication()
+            : base()
         {
             Initialize();
         }
 
         /// <summary>
-        /// Initializes a new instance of <see cref="CommandLineApplication"/>.
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
         /// </summary>
         /// <param name="console">The console implementation to use.</param>
         public CommandLineApplication(IConsole console)
@@ -42,11 +40,72 @@ namespace McMaster.Extensions.CommandLineUtils
         }
 
         /// <summary>
-        /// Initializes a new instance of <see cref="CommandLineApplication"/>.
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
+        /// </summary>
+        /// <param name="console">The console implementation to use.</param>
+        /// <param name="workingDirectory">The current working directory.</param>
+        public CommandLineApplication(IConsole console, string workingDirectory)
+            : base(console, workingDirectory)
+        {
+            Initialize();
+        }
+
+        /// <summary>
+        /// <para>
+        /// This constructor is obsolete and will be removed in a future version.
+        /// The recommended replacement is <see cref="CommandLineApplication{TModel}(IHelpTextGenerator, IConsole, string)" />
+        /// </para>
+        /// <para>
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
+        /// </para>
+        /// </summary>
+        /// <param name="helpTextGenerator">The help text generator to use.</param>
+        /// <param name="console">The console implementation to use.</param>
+        /// <param name="workingDirectory">The current working directory.</param>
+        public CommandLineApplication(IHelpTextGenerator helpTextGenerator, IConsole console, string workingDirectory)
+            : base(helpTextGenerator, console, workingDirectory)
+        {
+            Initialize();
+        }
+
+        /// <summary>
+        /// <para>
+        /// This constructor is obsolete and will be removed in a future version.
+        /// The recommended replacement is <see cref="CommandLineApplication{TModel}()" />.
+        /// See https://github.com/natemcmaster/CommandLineUtils/issues/339 for details.
+        /// </para>
+        /// <para>
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
+        /// </para>
+        /// </summary>
+        /// <param name="throwOnUnexpectedArg">Initial value for <see cref="CommandLineApplication.ThrowOnUnexpectedArgument"/>.</param>
+        [Obsolete("This constructor is obsolete and will be removed in a future version. " +
+            "The recommended replacement is CommandLineApplication<T>() and UnrecognizedArgumentHandling. " +
+            "See https://github.com/natemcmaster/CommandLineUtils/issues/339 for details.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public CommandLineApplication(bool throwOnUnexpectedArg)
+            : base(throwOnUnexpectedArg)
+        {
+            Initialize();
+        }
+
+        /// <summary>
+        /// <para>
+        /// This constructor is obsolete and will be removed in a future version.
+        /// The recommended replacement is <see cref="CommandLineApplication{TModel}(IConsole, string)" />
+        /// See https://github.com/natemcmaster/CommandLineUtils/issues/339 for details.
+        /// </para>
+        /// <para>
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
+        /// </para>
         /// </summary>
         /// <param name="console">The console implementation to use.</param>
         /// <param name="workingDirectory">The current working directory.</param>
         /// <param name="throwOnUnexpectedArg">Initial value for <see cref="CommandLineApplication.ThrowOnUnexpectedArgument"/>.</param>
+        [Obsolete("This constructor is obsolete and will be removed in a future version. " +
+            "The recommended replacement is CommandLineApplication<T>(IConsole, string) and UnrecognizedArgumentHandling. " +
+            "See https://github.com/natemcmaster/CommandLineUtils/issues/339 for details.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public CommandLineApplication(IConsole console, string workingDirectory, bool throwOnUnexpectedArg)
             : base(console, workingDirectory, throwOnUnexpectedArg)
         {
@@ -54,20 +113,31 @@ namespace McMaster.Extensions.CommandLineUtils
         }
 
         /// <summary>
-        /// Initializes a new instance of <see cref="CommandLineApplication"/>.
+        /// <para>
+        /// This constructor is obsolete and will be removed in a future version.
+        /// The recommended replacement is <see cref="CommandLineApplication{TModel}(IHelpTextGenerator, IConsole, string)" />
+        /// See https://github.com/natemcmaster/CommandLineUtils/issues/339 for details.
+        /// </para>
+        /// <para>
+        /// Initializes a new instance of <see cref="CommandLineApplication{TModel}"/>.
+        /// </para>
         /// </summary>
         /// <param name="helpTextGenerator">The help text generator to use.</param>
         /// <param name="console">The console implementation to use.</param>
         /// <param name="workingDirectory">The current working directory.</param>
         /// <param name="throwOnUnexpectedArg">Initial value for <see cref="CommandLineApplication.ThrowOnUnexpectedArgument"/>.</param>
+        [Obsolete("This constructor is obsolete and will be removed in a future version. " +
+            "The recommended replacement is CommandLineApplication<T>(IHelpTextGenerator, IConsole, string) and UnrecognizedArgumentHandling. " +
+            "See https://github.com/natemcmaster/CommandLineUtils/issues/339 for details.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public CommandLineApplication(IHelpTextGenerator helpTextGenerator, IConsole console, string workingDirectory, bool throwOnUnexpectedArg)
             : base(helpTextGenerator, console, workingDirectory, throwOnUnexpectedArg)
         {
             Initialize();
         }
 
-        internal CommandLineApplication(CommandLineApplication parent, string name, bool throwOnUnexpectedArg)
-            : base(parent, name, throwOnUnexpectedArg)
+        internal CommandLineApplication(CommandLineApplication parent, string name)
+            : base(parent, name)
         {
             Initialize();
         }