+++ test/CommandLineUtils.Tests/CommandLineApplicationExecutorTests.cs
@@ -232,7 +232,7 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
             Assert.Equal(200, await CommandLineApplication.ExecuteAsync<TaskReturnType>());
         }
 
-        [Command(ThrowOnUnexpectedArgument = false)]
+        [Command(UnrecognizedArgumentHandling = UnrecognizedArgumentHandling.StopParsingAndCollect)]
         private class ContextApp
         {
             public int OnExecute(CommandLineContext context)
@@ -313,7 +313,7 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
         private class Subcommand
         {
             public DisposableParentCommand Parent { get; }
-            
+
             public void OnExecute()
             {
                 Assert.NotNull(Parent);