+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 
 * Fix [#251] by [@mattnischan] - remove API that was marked as obsolete in 2.x releases
 * Fix [#294] by [@natemcmaster] - change dependencies on McMaster.Extensions.Hosting.CommandLine to just use Microsoft.Extensions.Hosting.Abstractions
+* Fix [#337] by [@natemcmaster] - removed .NET Standard 1.6 target from library
 
 [@mattnischan]: https://github.com/mattnischan
 [@natemcmaster]: https://github.com/natemcmaster
 
 [#251]: https://github.com/natemcmaster/CommandLineUtils/issues/251
 [#294]: https://github.com/natemcmaster/CommandLineUtils/issues/294
+[#337]: https://github.com/natemcmaster/CommandLineUtils/issues/337
 
 ## [v2.5.1](https://github.com/natemcmaster/CommandLineUtils/compare/v2.5.0...v2.5.1)